backend/foodgram/requirements.txt
Django~=2.2.19
djangorestframework~=3.12.4
django-stubs==1.12.0
django-stubs-ext==0.5.0
drf-extra-fields==3.4.0
Pillow==9.2.0
djoser==2.1.0
reportlab==3.6.11
django-filter==21.1
